add casts to force correct division

+++ pandas/_libs/tslibs/timedeltas.pyx
@@ -865,7 +865,8 @@ cdef class _Timedelta(timedelta):
         """
         # GH 31043
         # Round down to microseconds to avoid confusing tzinfo.utcoffset
-        return (self.value - self.value % 1000) / 1000000000
+        # use casts to make sure precision isn't lost
+        return np.float64(int(self.value - self.value % 1000) / 1000000000)
 
     def view(self, dtype):
         """
